requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
aiofiles==23.2.1
python-dotenv==1.0.1
psutil==5.9.8